feat(vote): refactor commands into files and dirs
